Update streaming libs
