+++ lib/streaming/peer-connection.js
@@ -118,13 +118,6 @@ export class WeriftPeerConnection extends Subscribed {
         return offer
     }
 
-    async createAnswer(offer) {
-        await this.pc.setRemoteDescription(offer)
-        const answer = await this.pc.createAnswer()
-        await this.pc.setLocalDescription(answer)
-        return answer
-    }
-
     async acceptAnswer(answer) {
         await this.pc.setRemoteDescription(answer)
     }
