+++ lib/streaming/webrtc-connection.js
@@ -2,30 +2,73 @@
 // to native Javascript with custom logging for ring-mqtt and some unused code removed.
 // Much thanks to @dgreif for the original code which is the basis for this work.
 
-import WebSocket from 'ws'
 import { parentPort } from 'worker_threads'
-import { firstValueFrom, interval, ReplaySubject } from 'rxjs'
-import { StreamingConnectionBase } from './streaming-connection-base.js'
+import { WebSocket } from 'ws'
+import { firstValueFrom, fromEvent, interval, ReplaySubject } from 'rxjs'
+import { concatMap, take } from 'rxjs/operators'
 import crypto from 'crypto'
+import { WeriftPeerConnection } from './peer-connection.js'
+import { Subscribed } from './subscribed.js'
 
-export class WebrtcConnection extends StreamingConnectionBase {
+export class WebrtcConnection extends Subscribed {
     constructor(ticket, camera) {
-        super(
-            new WebSocket(`wss://api.prod.signalling.ring.devices.a2z.com:443/ws?api_version=4.0&auth_type=ring_solutions&client_id=ring_site-${crypto.randomUUID()}&token=${ticket}`, {
+        super()
+        this.ws = new WebSocket(
+            `wss://api.prod.signalling.ring.devices.a2z.com:443/ws?api_version=4.0&auth_type=ring_solutions&client_id=ring_site-${crypto.randomUUID()}&token=${ticket}`,
+            {
                 headers: {
                     // This must exist or the socket will close immediately but content does not seem to matter
                     'User-Agent': 'android:com.ringapp'
                 }
-            })
-
+            }
         )
         this.camera = camera
         this.onSessionId = new ReplaySubject(1)
         this.onOfferSent = new ReplaySubject(1)
         this.sessionId = null
         this.dialogId = crypto.randomUUID()
+        this.onCameraConnected = new ReplaySubject(1)
+        this.onCallAnswered = new ReplaySubject(1)
+        this.onCallEnded = new ReplaySubject(1)
+        this.onMessage = new ReplaySubject()
+        this.hasEnded = false
+        const pc = new WeriftPeerConnection()
+        this.pc = pc
+        this.onAudioRtp = pc.onAudioRtp
+        this.onVideoRtp = pc.onVideoRtp
+        this.onWsOpen = fromEvent(this.ws, 'open')
+        const onMessage = fromEvent(this.ws, 'message')
+        const onError = fromEvent(this.ws, 'error')
+        const onClose = fromEvent(this.ws, 'close')
 
         this.addSubscriptions(
+            onMessage.pipe(concatMap((event) => {
+                const message = JSON.parse(event.data)
+                this.onMessage.next(message)
+                return this.handleMessage(message)
+            })).subscribe(),
+
+            onError.subscribe((error) => {
+                parentPort.postMessage({type: 'log_error', data: error})
+                this.callEnded()
+            }),
+
+            onClose.subscribe(() => {
+                this.callEnded()
+            }),
+
+            this.pc.onConnectionState.subscribe((state) => {
+                if (state === 'failed') {
+                    parentPort.postMessage({type: 'log_error', data: 'WebRTC peer connection failed'})
+                    this.callEnded()
+                }
+
+                if (state === 'closed') {
+                    parentPort.postMessage({type: 'log_info', data: 'WebRTC peer connection closed'})
+                    this.callEnded()
+                }
+            }),
+
             this.onWsOpen.subscribe(() => {
                 parentPort.postMessage({type: 'log_info', data: 'Websocket signaling for WebRTC session connected successfully'})
                 this.initiateCall().catch((error) => {
@@ -109,8 +152,13 @@ export class WebrtcConnection extends StreamingConnectionBase {
                 return
             case 'notification':
                 const { text } = message.body
-                if (text === 'PeerConnectionState::kConnecting' ||
-                    text === 'PeerConnectionState::kConnected') {
+                if (text === 'camera_connected') {
+                    this.onCameraConnected.next()
+                    return
+                } else if (
+                    text === 'PeerConnectionState::kConnecting' ||
+                    text === 'PeerConnectionState::kConnected'
+                ) {
                     return
                 }
                 break
@@ -121,12 +169,12 @@ export class WebrtcConnection extends StreamingConnectionBase {
     }
 
     sendSessionMessage(method, body = {}) {
-        const sendSessionMessage = () => {
+        const sendSessionMessage = (sessionId) => {
             const message = {
                 body: {
                     ...body,
                     doorbot_id: this.camera.id,
-                    session_id: this.sessionId,
+                    session_id: sessionId,
                 },
                 dialog_id: this.dialogId,
                 method
@@ -136,14 +184,55 @@ export class WebrtcConnection extends StreamingConnectionBase {
         if (this.sessionId) {
             // Send immediately if we already have a session id
             // This is needed to send `close` before closing the websocket
-            sendSessionMessage()
+            sendSessionMessage(this.sessionId)
+        } else {
+            this.addSubscriptions(
+                this.onSessionId.pipe(take(1)).subscribe(sendSessionMessage)
+            )
         }
-        else {
-            firstValueFrom(this.onSessionId)
-                .then(sendSessionMessage)
-                .catch((e) => {
-                    // debug(e)
-                })
+    }
+
+    sendMessage(message) {
+        if (this.hasEnded) {
+            return
         }
+        this.ws.send(JSON.stringify(message))
+    }
+
+    activate() {
+        // the activate_session message is required to keep the stream alive longer than 70 seconds
+        this.sendSessionMessage('activate_session')
+        this.sendSessionMessage('stream_options', {
+            audio_enabled: true,
+            video_enabled: true,
+        })
+    }
+
+    callEnded() {
+        if (this.hasEnded) {
+            return
+        }
+        try {
+            this.sendMessage({
+                reason: { code: 0, text: '' },
+                method: 'close',
+            })
+            this.ws.close()
+        }
+        catch (_) {
+            // ignore any errors since we are stopping the call
+        }
+        this.hasEnded = true
+        this.unsubscribe()
+        this.onCallEnded.next()
+        this.pc.close()
+    }
+
+    stop() {
+        this.callEnded()
+    }
+
+    requestKeyFrame() {
+        this.pc.requestKeyFrame?.()
     }
 }
